Safari-609.1.11.1/LayoutTests/platform/wpe/fast/canvas/webgl/index-validation-with-subsequent-draws-expected.txt

CONSOLE MESSAGE: line 43: TypeError: null is not an object (evaluating 'ext.drawElementsInstancedANGLE')
